Fix all ruff lint violations and apply formatting
Resolve 3,377 ruff violations across the codebase by applying auto-fixes,
unsafe fixes, and manual corrections. Expand ruff config with new rule
sets (naming, comprehensions, pathlib, print-detection, return,
type-checking, pygrep-hooks, perflint) and add sensible ignores. All
2,755 unit tests pass.

Key changes:
- Convert for-append loops to list comprehensions/extend (PERF401)
- Merge collapsible nested ifs (SIM102)
- Replace open() with Path.open() (PTH123)
- Move typing-only imports behind TYPE_CHECKING (TC001/TC003)
- Fix broken imports from TC003 unsafe moves (F821)
- Add from __future__ import annotations where needed
- Apply ruff formatter to 391 files
- Add per-file ignores for examples, tests, and load-balancing module

+++ tests/unit/sketching/test_count_min_sketch.py
@@ -1,12 +1,12 @@
 """Tests for Count-Min Sketch."""
 
-from collections import Counter
 import math
+from collections import Counter
 
 import pytest
 
-from happysimulator.sketching import CountMinSketch, FrequencyEstimate
 from happysimulator.distributions import ZipfDistribution
+from happysimulator.sketching import CountMinSketch, FrequencyEstimate
 
 
 class TestCountMinSketchCreation:
@@ -235,7 +235,7 @@ class TestCountMinSketchErrorBounds:
         assert estimate.item == 42
         assert estimate.count >= 100
         # Error bound is epsilon * total_count
-        expected_error = int(math.ceil(cms.epsilon * cms.item_count))
+        expected_error = math.ceil(cms.epsilon * cms.item_count)
         assert estimate.error == expected_error
